requirements.txt
discord.py==0.16.12
beautifulsoup4==4.6.0
requests==2.18.4
lxml==4.1.1